requirements.txt
flask
opencv-python
trilobot
python-dotenv 